Update signatures.md
Initially had a question about this sentence: "Note that the hosted version of Fluree does not allow you to sign queries, because fdb-api-open is set to true for all hosted accounts, so a signed query would be ignored regardless". Question was - 'Why is fdb-api-open set to true for all hosted accounts?'. Per Andrew, this is not the case anymore, so we should delete this sentence from the documentation.